Introduzione del punteggio finale nella ResultsActivity e introduzione del CustomAdapter per la gestione della listView (ancora da concludere)

+++ app/src/main/java/it/giuliatesta/udrive/DataProcessor.java
@@ -149,7 +149,7 @@ public class DataProcessor {
         @param max valore massimo
         @return valore medio
      */
-    private double getMediumValue(int max, int min) {
+    private double getMediumValue(double max, double min) {
         return (max + min)/2;
     }
